telegram_bot/requirements.txt
aiogram==2.11.2
aiohttp==3.7.3
aioredis==1.3.1
async-timeout==3.0.1
asyncpg==0.21.0
attrs==20.3.0
Babel==2.9.0
certifi==2020.11.8
chardet==3.0.4
gino==1.0.1
hiredis==1.1.0
idna==2.10
importlib-metadata==1.7.0
multidict==5.0.2
python-dotenv==0.15.0
pytz==2020.4
SQLAlchemy==1.3.20
typing-extensions==3.7.4.3
yarl==1.6.3
zipp==3.4.0
